moneypy/requirements.txt
requests==2.32.3
pandas==2.0.3
pyyaml==6.0.2
